Move yolo_vanilla back to yolo.py
